Fix show ip/v6 route summary non-multi-asic platform to interact with FRR directly (#1306)
Changed the handling of "show ip/v6 route xxxx" for non-multi-asic platforms to be handled by FRR completely instead of using the common code added to support multi-asic platform. This is because there are some options such as summary and tables are not yet supported by multi-asic platform and we should not penalize non-multi-asic platform by sharing the same common code.
Due to this change, had to modify/add additional unit tests to ensure good test coverage of > 95%.
Took the opportunity to refactor the expected test data output strings into a standalone file to make the test code easier to read.