route_check: Fix hanging & logging level (#1520)
Fix for hanging:
Enforce max len on string to be printed to stdout.
With huge mismatch, the message to print can be too long to hang python print statement.

Lower logging level:
The log for "details of route mismatch" is lowered to WARNING. The caller of the route_check may log error, as it sees it fit. The failure of route_check is indicated to caller via exit code.

How to verify it
Simulate error (route-mismatch), and verify
a) message len does not exceed set MAX
b) log level is WARNING