Copy missing values from INIT_CFG to config_db during db_migration (#1522)
FEATURE table is a part of 201911 image database, but absent in 201811 images. The warm upgrade from 201811 image to 201911 image fails as the feature table is missing in the DB and multiple errors are seen in syslog.

This PR fixes DB_MIGRATOR code to update missing init_cfg tables into config_db.