[acl_loader]: add iptype match to the rules for dataplane acl
dataplane acl has v4 and v6 type. in case the rule does not
specify the iptype, the acl_loader will automatically add
the match for the iptype based on the table type.

for l3 table, it will add ethertype = 0x800
for l3v6 table, it will add iptype = ipv6any

Signed-off-by: Guohan Lu <lguohan@gmail.com>