[201911][acl] Use a list instead of a comma-separated string for ACL port list (#1576)
How I did it
I discovered that using the @ syntax to insert the port list as a comma-separated list/array actually causes the port list to be deleted when the same table is config add-ed multiple times. Using a list works. (We still need to determine the root cause for why the @ syntax fails).

How to verify it
Updated the tests and ran the command locally to verify the fix. Still working on adding a new regression test case for this.