Issue #22407: ConfigReload fails when RADIUS statistics is enabled (#3860)
What I did
Enabling RADIUS statistics adds config entries which are not valid in YANG. The option adds the True/False boolean values instead of the 'true'/'false' booleans that are defined in the YANG model, which causes failures when trying to reload/replace the config.

Fixes: #22407

How I did it
Modified relevant config command to use the correct values.

How to verify it
Manually verified.

I added a test case for config radius statistics command. Note that this test follows the pattern in this file, which is not actually exercising the correct codepath (see #3754 for details).