`vnet_route_check.py` should not report VNET routes in APP DB but not in STATE DB and ASIC DB as mismatches (#3990)
What I did
Changed vnet_route_check.py so that when a VNET route is missing in STATE DB, then it is treated as an inactive route. This means that when the script is called without -a and --all options, if a route in APP DB is missing in both STATE DB and ASIC DB, it will not be reported as a mismatch.

How I did it
Changed the logic in vnet_route_check.py.
Added a new test and modified an existing test in vnet_route_check_test.py to verify the desired behavior.
How to verify it
Run the tests in vnet_route_check_test.py.